Move setupTests.tsx to another folder
