+++ .eslintrc.json
@@ -10,6 +10,8 @@
   "parserOptions": { "project": ["./tsconfig.json"] },
   "plugins": ["@typescript-eslint"],
   "rules": {
+    "@typescript-eslint/no-empty-function": "off",
+
     "no-empty": ["error", { "allowEmptyCatch": true }],
     "@typescript-eslint/no-var-requires": 0,
     "@typescript-eslint/no-explicit-any": "off",
